# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Proyecto Final PAEP
  version: 1.0.0
paths:
  /users:
    get:
      tags:
      - TSP
      summary: Get users
      responses:
        200:
          description: Sorted users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Save a new user
      tags:
      - TSP
      responses:
        400:
          description: Some field is wrong/missing
          content:
            application/json:
              schema:
                type: string
        200:
          description: Creation of a new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
  /users/${username}:
    get:
      tags:
      - TSP
      parameters:
      - name: username
        in: path
        required: true
      summary: Get a specific username
      responses:
        200:
          $ref: '#/components/responses/200User'
  /users/${username}/locations:
    get:
      tags:
      - TSP
      parameters:
      - name: username
        in: path
        required: true
      summary: Get the locations of a specific user
      responses:
        200:
          $ref: '#/components/responses/200UserLocations'
  /users/:username/addLocation/:name:
    put:
      tags:
      - TSP
      parameters:
      - name: username
        in: path
        required: true
      - name: name
        in: path
        required: true
      - name: filterBy
        in: query
      summary: Adds a location to a specific user
      responses:
        200:
          $ref: '#/components/responses/200User'
components:
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxlength: 30
        lastName:
          type: string
          minLength: 3
          maxlength: 30
        username:
          type: string
          minLength: 6
          maxlength: 30
        locations:
          $ref: '#/components/schemas/LocationsArray'
    LocationsArray:
      type: array
      items:
        type: integer
    Location:
      type: object
      properties:
        longitude:
          type: integer
          format: float
          maximum: 90
          minimum: -90
        latitude:
          type: integer
          format: float
          maximum: 180
          minimum: -180
        name:
          type: string
  responses:
    200User:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'
    200UserLocations:
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Location'